Split FastAPI routes into dedicated modules
