+++ leropa/web/routes/__init__.py
@@ -1,0 +1,3 @@
+"""Individual route modules for the web application."""
+
+from __future__ import annotations
